refactor(xhci): extract control_transfer to transfers module
Move the control_transfer method out of XhciContext into a dedicated transfers module to improve code organization and separation of concerns. Update the visibility of driver_ctx and deferred_free_list to pub(super) so they remain accessible from child modules.
