+++ nitrogen/src/usb/xhci/mod.rs
@@ -4,3 +4,4 @@ pub mod interrupt;
 pub mod port;
 pub mod register;
 pub mod ring;
+pub mod transfer;
